.ipynb_checkpoints/requirements-checkpoint.txt
accelerate==0.21.0
einops==0.6.1
ema_pytorch==0.2.3
gym==0.26.2
imageio==2.31.1
matplotlib==3.6.2
numpy==1.22.2
opencv==4.6.0
packaging==23.1
pandas==1.5.2
Pillow==9.2.0
Pillow==10.0.0
pytorch_fid==0.3.0
PyYAML==6.0.1
PyYAML==6.0.1
torch==1.14.0a0+410ce96
torchvision==0.15.0a0
tqdm==4.64.1
ipywidgets